requirements.txt
black==24.4.2
click==8.1.7
Markdown==3.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyYAML==6.0.1
